2-strlen.c updated

+++ 0x05-pointers_arrays_strings/2-strlen.c
@@ -2,12 +2,13 @@
 #include <stdio.h>
 
 /**
- * _strlen - a function that returns the length of a string.
+ * _strlen - returns the length of a string.
  * @str: The string to get the length of
  * Return: The length of @str
  */
 
-int _strlen(const char *str);
+size_t _strlen(const char *str);
+
 {
 	size_t length = 0;
